Repackage Clerk-related services to dedicated module
