+++ api/src/main/java/com/example/echo_api/modules/clerk/service/ClerkSdkService.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.service.clerk.sdk;
+package com.example.echo_api.modules.clerk.service;
 
 import java.util.List;
