+++ api/src/main/java/com/example/echo_api/modules/clerk/service/ClerkWebhookServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.service.clerk.webhook;
+package com.example.echo_api.modules.clerk.service;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
@@ -7,7 +7,6 @@ import com.clerk.backend_api.Clerk;
 import com.example.echo_api.exception.custom.badrequest.DeserializationException;
 import com.example.echo_api.exception.custom.unauthorised.WebhookVerificationException;
 import com.example.echo_api.persistence.dto.request.clerk.webhook.ClerkWebhookEvent;
-import com.example.echo_api.service.clerk.sync.ClerkSyncService;
 import com.example.echo_api.util.Utils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.svix.Webhook;
